+++ speedlib/src/main/java/com/tools/speedlib/listener/SpeedListener.java
@@ -5,7 +5,7 @@ package com.tools.speedlib.listener;
  * Created by wong on 17-3-27.
  */
 public interface SpeedListener {
-    void speeding(double downSpeed, double upSpeed);
+    void speeding(long downSpeed, long upSpeed);
 
-    void finishSpeed(double finalDownSpeed, double finalUpSpeed);
+    void finishSpeed(long finalDownSpeed, long finalUpSpeed);
 }
